BHY2: Refactor sensors
Sensors can be instantiated as object of type: SensorOrientation or SensorXYZ.
A sensor ID should be passed to the constructor of the object.
After being configured, the sensor data is automatically filled by the library through the update function.
When a sensor is read (by accessing the objects methods), its last value received from the BHY is returned.

A Sensor Manager is responsible for handling the sensors instantiated.
A sensor is subscribed to the sensor manager during its construction. It is unsubscribed during its destruction.
TODO: Allow dyanimic subscription/unsubscription depending on the last configuration of the sensor. If it is off (rate = 0) then unsubscribe.